feat: Move `countRecord` method to `CountUtil`.
