+++ src/main/java/org/spin/base/util/RecordUtil.java
@@ -50,6 +50,7 @@ import org.compiere.util.Util;
 import org.spin.backend.grpc.common.Entity;
 import org.spin.backend.grpc.common.ListEntitiesResponse;
 import org.spin.backend.grpc.common.Value;
+import org.spin.base.db.FromUtil;
 import org.spin.model.MADAttachmentReference;
 import org.spin.util.AttachmentUtil;
 
@@ -473,7 +474,7 @@ public class RecordUtil {
 				String keyColumn = table.getTableName() + "_ID";
 				String translationTableName = tableTranslation.getTableName();
 				// only table, tableName tableName, tableName AS tableName
-				String tableWithAliases = getPatternTableName(translationTableName, null);
+				String tableWithAliases = FromUtil.getPatternTableName(translationTableName, null);
 
 				String joinPattern = "JOIN(\\s+)" 
 					+ "(" + tableWithAliases + ")"
@@ -514,7 +515,7 @@ public class RecordUtil {
 			}
 
 			// tableName tableName, tableName AS tableName
-			String patternAlias = getPatternTableName(table.getTableName(), tableAlias);
+			String patternAlias = FromUtil.getPatternTableName(table.getTableName(), tableAlias);
 
 			String fromWherePattern = "\\s+(FROM)\\s+(" + patternAlias + ")" + "\\s+(WHERE)";
 			Matcher matcher = Pattern.compile(
@@ -558,83 +559,6 @@ public class RecordUtil {
 	}
 
 
-	/**
-	 * Get regex pattern to match with table name and/or table alias
-	 * @param tableName
-	 * @param tableNameAlias
-	 * @return
-	 */
-	public static String getPatternTableName(String tableName, String tableNameAlias) {
-		// tableName tableName, tableName AS tableName
-		String patternTableWithAliases = tableName + "\\s+" + tableName + "|" + tableName + "\\s+AS\\s+" + tableName;
-		if (!Util.isEmpty(tableNameAlias, true) && !tableName.equals(tableNameAlias)) {
-			// tableName tableAlias, tableName AS tableAlias
-			patternTableWithAliases += "|" + tableName + "\\s+" + tableNameAlias + "|" + tableName + "\\s+AS\\s+" + tableNameAlias;
-		}
-		// only table on last position
-		patternTableWithAliases += "|" + tableName;
-
-		return patternTableWithAliases;
-	}
-
-	/**
-	 * Count records
-	 * @param sql
-	 * @param tableName
-	 * @param parameters
-	 * @return
-	 */
-	public static int countRecords(String sql, String tableName, List<Object> parameters) {
-		return countRecords(sql, tableName, null, parameters);
-	}
-
-	/**
-	 * Count records
-	 * @param sql
-	 * @param tableName
-	 * @param tableNameAlias
-	 * @param parameters
-	 * @return
-	 */
-	public static int countRecords(String sql, String tableName, String tableNameAlias, List<Object> parameters) {
-		// tableName tableName, tableName AS tableName
-		String tableWithAliases = getPatternTableName(tableName, tableNameAlias);
-
-		Matcher matcherFrom = Pattern.compile(
-			"\\s+(FROM)\\s+(" + tableWithAliases + ")\\s+(WHERE|(LEFT|INNER|RIGHT)\\s+JOIN)",
-			Pattern.CASE_INSENSITIVE | Pattern.DOTALL
-		)
-		.matcher(sql);
-
-		List<MatchResult> fromWhereParts = matcherFrom.results()
-			.collect(Collectors.toList());
-		int positionFrom = -1;
-		if (fromWhereParts != null && fromWhereParts.size() > 0) {
-			// MatchResult lastFrom = fromWhereParts.get(fromWhereParts.size() - 1);
-			MatchResult lastFrom = fromWhereParts.get(0);
-			positionFrom = lastFrom.start();
-		} else {
-			return 0;
-		}
-
-		String queryCount = "SELECT COUNT(*) " + sql.substring(positionFrom, sql.length());
-
-		// remove order by clause
-		Matcher matcherOrderBy = Pattern.compile(
-			"\\s+(ORDER BY)\\s+",
-			Pattern.CASE_INSENSITIVE | Pattern.DOTALL
-		).matcher(queryCount);
-		if(matcherOrderBy.find()) {
-			int positionOrderBy = matcherOrderBy.start();
-			queryCount = queryCount.substring(0, positionOrderBy);
-		}
-
-		if (parameters == null) {
-			parameters = new ArrayList<Object>();
-		}
-
-		return DB.getSQLValueEx(null, queryCount, parameters);
-	}
 
 	/**
 	 * Get Query with limit
