+++ src/main/java/org/spin/grpc/service/PayrollActionNoticeServiceImplementation.java
@@ -58,6 +58,7 @@ import org.spin.backend.grpc.form.ListPayrollMovementsRequest;
 import org.spin.backend.grpc.form.ListPayrollProcessRequest;
 import org.spin.backend.grpc.form.PayrollActionNoticeGrpc.PayrollActionNoticeImplBase;
 import org.spin.backend.grpc.form.SavePayrollMovementRequest;
+import org.spin.base.db.CountUtil;
 import org.spin.base.util.LookupUtil;
 import org.spin.base.util.RecordUtil;
 import org.spin.base.util.SessionManager;
@@ -150,7 +151,7 @@ public class PayrollActionNoticeServiceImplementation extends PayrollActionNotic
 		}
 		
 		//	Count records
-		int count = RecordUtil.countRecords(sql, MHRProcess.Table_Name, parameters);
+		int count = CountUtil.countRecords(sql, MHRProcess.Table_Name, parameters);
 		lookupsList.setRecordCount(count);
 
 		//	Set page token
@@ -273,7 +274,7 @@ public class PayrollActionNoticeServiceImplementation extends PayrollActionNotic
 		}
 
 		//	Count records
-		int count = RecordUtil.countRecords(sql, MHRProcess.Table_Name, parameters);
+		int count = CountUtil.countRecords(sql, MHRProcess.Table_Name, parameters);
 		lookupsList.setRecordCount(count);
 
 		//	Set page token
@@ -418,7 +419,7 @@ public class PayrollActionNoticeServiceImplementation extends PayrollActionNotic
 		}
 
 		//	Count records
-		int count = RecordUtil.countRecords(sql, MHRProcess.Table_Name, parameters);
+		int count = CountUtil.countRecords(sql, MHRProcess.Table_Name, parameters);
 		listLookups.setRecordCount(count);
 
 		//	Set page token
